Always uninitialize nodes when doing connections. Guard some graph functions with @synchronized()
Changing sample rate doesn't rebuild the graph; now requires a restart. (Can't reliably be done with encoders/streams active)

+++ CocoaSplit/CAMultiAudio/CAMultiAudioEngine.h
@@ -43,7 +43,7 @@
 
 @property (strong) CAMultiAudioDownmixer *encodeMixer;
 //@property (strong) CSAacEncoder *encoder;
-@property (assign) UInt32 sampleRate;
+@property (assign) double sampleRate;
 @property (assign) int audioBitrate;
 @property (assign) double audio_adjust;
 @property (strong) NSArray *audioOutputs;
